+++ node-src/graphQL/Tag/schema.js
@@ -7,7 +7,7 @@ export const type = `
     userId: String
     backgroundColor: String
     textColor: String
-    timestamp: Int
+    timestamp: Float
   }
 
   type TagQueryResults {
@@ -40,7 +40,7 @@ export const type = `
     userId: String
     backgroundColor: String
     textColor: String
-    timestamp: Int
+    timestamp: Float
   }
 
   input TagMutationInput {
@@ -49,7 +49,7 @@ export const type = `
     userId: String
     backgroundColor: String
     textColor: String
-    timestamp: Int
+    timestamp: Float
     timestamp_INC: Int
     timestamp_DEC: Int
   }
@@ -103,13 +103,13 @@ export const type = `
     textColor: String
     textColor_ne: String
     textColor_in: [String]
-    timestamp_lt: Int
-    timestamp_lte: Int
-    timestamp_gt: Int
-    timestamp_gte: Int
-    timestamp: Int
-    timestamp_ne: Int
-    timestamp_in: [Int]
+    timestamp_lt: Float
+    timestamp_lte: Float
+    timestamp_gt: Float
+    timestamp_gte: Float
+    timestamp: Float
+    timestamp_ne: Float
+    timestamp_in: [Float]
     OR: [TagFilters]
   }
   
@@ -185,13 +185,13 @@ export const query = `
     textColor: String,
     textColor_ne: String,
     textColor_in: [String],
-    timestamp_lt: Int,
-    timestamp_lte: Int,
-    timestamp_gt: Int,
-    timestamp_gte: Int,
-    timestamp: Int,
-    timestamp_ne: Int,
-    timestamp_in: [Int],
+    timestamp_lt: Float,
+    timestamp_lte: Float,
+    timestamp_gt: Float,
+    timestamp_gte: Float,
+    timestamp: Float,
+    timestamp_ne: Float,
+    timestamp_in: [Float],
     OR: [TagFilters],
     SORT: TagSort,
     SORTS: [TagSort],
